% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dplyr.R
\name{filter.Shennong}
\alias{filter.Shennong}
\title{Filter samples in a Shennong object using dplyr syntax}
\usage{
\method{filter}{Shennong}(.data, ..., .by = NULL, .preserve = FALSE)
}
\arguments{
\item{.data}{A `Shennong` object.}

\item{...}{Filtering conditions passed to `dplyr::filter()`.}

\item{.by}{Optional grouping variable(s).}

\item{.preserve}{Whether to preserve the grouping structure.}
}
\value{
A filtered `Shennong` object.
}
\description{
Allows tidyverse-style filtering of sample metadata (`colData`) in a Shennong object.
This behaves like `dplyr::filter()`, and subsets the object to retain only the samples
matching the specified conditions.
}
\examples{
\dontrun{
filter(so, stage == "Stage I", gender == "Female")
}

}
