% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ggplot2.R
\name{ggplot.Shennong}
\alias{ggplot.Shennong}
\title{ggplot method for Shennong object}
\usage{
\method{ggplot}{Shennong}(data, mapping = aes(), ..., environment = parent.frame(), vars = NULL)
}
\arguments{
\item{data}{A Shennong object.}

\item{mapping}{Aesthetic mapping (e.g., \code{aes(x = ..., y = ..., color = ...)}).}

\item{...}{Additional arguments passed to \code{ggplot()}.}

\item{environment}{The environment in which to evaluate the \code{mapping}. Usually left as default.}

\item{vars}{Optional character vector of variables to fetch using \code{sn_fetch_data()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Allows tidy ggplot2 visualization using sample-level metadata or expression values.
}
