% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-reduction.R,
%   R/show-method.R
\docType{class}
\name{Shennong-class}
\alias{Shennong-class}
\alias{Shennong}
\alias{[[,Shennong,character,ANY-method}
\alias{[[<-,Shennong,character,missing,ShennongReduction-method}
\alias{show,Shennong-method}
\title{Show method for Shennong objects}
\usage{
\S4method{[[}{Shennong,character,ANY}(x, i, j, ...)

\S4method{[[}{Shennong,character,missing,ShennongReduction}(x, i, j, ...) <- value

\S4method{show}{Shennong}(object)
}
\arguments{
\item{x}{A Shennong object}

\item{i}{Name of the reduction to assign}

\item{j}{(Ignored) Required for method signature but not used.}

\item{...}{Additional arguments (ignored)}

\item{value}{A ShennongReduction object}

\item{object}{A Shennong object.}
}
\description{
Displays summary information about the Shennong object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{x[[i}: Assign a reduction to the object by name

\item \code{`[[`(x = Shennong, i = character, j = missing) <- value}: Assign a reduction to the object by name

}}
\section{Slots}{

\describe{
\item{\code{organism}}{Character string specifying the organism (e.g., "human").}

\item{\code{active_assay}}{Character string indicating the name of the currently active
assay within the \code{ExperimentList}.}

\item{\code{active_ident}}{A factor representing the current identity (e.g., cluster
or condition) for each sample in the object's primary \code{colData}.
The length must match \code{nrow(colData(object))}.}

\item{\code{reductions}}{A list storing dimensional reduction information (e.g., PCA,
tSNE, UMAP). Typically, each element is a matrix where rows correspond to
samples. It's recommended to store reductions associated with specific assays
or the primary sample set.}

\item{\code{project_name}}{Character string for the name of the project.}

\item{\code{misc}}{A list for storing miscellaneous data associated with the object.}

\item{\code{version}}{The version of the ShennongObject package used to create the object.}

\item{\code{commands}}{A list to store analysis commands history (audit trail).}

\item{\code{tools}}{A list to store results generated by various tools or functions.}
}}

