% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ShennongCommand-class}
\alias{ShennongCommand-class}
\title{The ShennongCommand Class}
\description{
Used to log function calls, parameters, timestamps, and assay context
for tracking operations within a Shennong object
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the command}

\item{\code{time_stamp}}{POSIXct timestamp of command execution}

\item{\code{assay_used}}{Name of assay used (or NA if not set)}

\item{\code{call_string}}{Full function call string}

\item{\code{params}}{List of parameters passed to the command}
}}

\concept{command}
