% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShennongObject-package.R
\docType{package}
\name{ShennongObject-package}
\alias{ShennongObject}
\alias{ShennongObject-package}
\title{ShennongObject: Modern Data Structures for Multi-Omics Analysis}
\description{
Provides a tidy, extensible, and modular S4 data infrastructure for organizing and analyzing multi-omics and single-cell datasets in R. The ShennongObject class extends MultiAssayExperiment to support streamlined integration of assays, flexible metadata management, dimensionality reductions, visualization tools, and analysis provenance tracking. Designed for compatibility with tidyverse-style workflows and modern bioinformatics pipelines.
}
\author{
\strong{Maintainer}: Songqi Duan \email{duan@songqi.org} (\href{https://orcid.org/0000-0002-0822-5883}{ORCID})

}
\keyword{internal}
