% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ShennongTool-class}
\alias{ShennongTool-class}
\title{The ShennongTool Class}
\description{
Stores the results generated by a computational tool or algorithm
(e.g., UCell, GSVA, AUCell) applied to data in a \code{Shennong} object.
}
\section{Slots}{

\describe{
\item{\code{method}}{Character. The name of the tool or method used (e.g., "UCell").}

\item{\code{version}}{Character. Version of the tool used.}

\item{\code{result}}{A list or matrix storing the primary result.}

\item{\code{params}}{A list of parameters used when running the tool.}

\item{\code{timestamp}}{POSIXct. The time when the tool was run.}

\item{\code{status}}{Character. Status of the run ("success", "error", etc.).}

\item{\code{citation}}{Character. Optional reference or citation for the tool.}

\item{\code{misc}}{A list of additional metadata.}
}}

