% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\name{show,ShennongReduction-method}
\alias{show,ShennongReduction-method}
\title{Show method for ShennongReduction}
\usage{
\S4method{show}{ShennongReduction}(object)
}
\arguments{
\item{object}{A ShennongReduction object.}
}
\description{
Show method for ShennongReduction
}
