% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\name{show,ShennongCommand-method}
\alias{show,ShennongCommand-method}
\title{Show method for ShennongCommand}
\usage{
\S4method{show}{ShennongCommand}(object)
}
\arguments{
\item{object}{A ShennongCommand object}
}
\value{
Invisibly returns the object.
}
\description{
Displays key information for a ShennongCommand object.
}
