% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_active_assay}
\alias{sn_active_assay}
\alias{sn_active_assay<-}
\alias{sn_active_assay,Shennong-method}
\alias{sn_active_assay<-,Shennong-method}
\title{Get Default Assay}
\usage{
sn_active_assay(object, ...)

sn_active_assay(object, ...) <- value

\S4method{sn_active_assay}{Shennong}(object, ...)

\S4method{sn_active_assay}{Shennong}(object, ...) <- value
}
\arguments{
\item{object}{A Shennong object.}

\item{...}{Additional arguments.}

\item{value}{The assay name to set as active.}
}
\description{
Get Default Assay

Set Default Assay<-
}
