% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_active_ident}
\alias{sn_active_ident}
\alias{sn_active_ident<-}
\alias{sn_active_ident,Shennong-method}
\alias{sn_active_ident<-,Shennong-method}
\title{Get Sample Identities}
\usage{
sn_active_ident(object, ...)

sn_active_ident(object, ...) <- value

\S4method{sn_active_ident}{Shennong}(object, ...)

\S4method{sn_active_ident}{Shennong}(object, samples = NULL, drop = FALSE, replace = FALSE, ...) <- value
}
\arguments{
\item{object}{A Shennong object.}

\item{...}{Additional arguments.}

\item{value}{Identities to set (factor, vector, or metadata column name).}

\item{samples}{Optional: Samples to set identities for.}

\item{drop}{Logical; drop unused identities.}

\item{replace}{Logical; replace existing identities.}
}
\description{
Get Sample Identities

Set Sample Identities<-
}
