% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_active_layer}
\alias{sn_active_layer}
\alias{sn_active_layer<-}
\alias{sn_active_layer,Shennong-method}
\alias{sn_active_layer<-,Shennong-method}
\title{Get the active layer}
\usage{
sn_active_layer(object, assay = NULL, ...)

sn_active_layer(object, assay = NULL, ...) <- value

\S4method{sn_active_layer}{Shennong}(object, assay = NULL, ...)

\S4method{sn_active_layer}{Shennong}(object, assay = NULL, ...) <- value
}
\arguments{
\item{object}{A Shennong object.}

\item{assay}{Optional: Assay name.}

\item{...}{Ignored.}

\item{value}{A character string of the layer name.}
}
\value{
A character scalar representing the current active layer.

The modified Shennong object.
}
\description{
Get the active layer

Set the active layer
}
