% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-reduction.R
\name{sn_active_reduction}
\alias{sn_active_reduction}
\alias{sn_active_reduction<-}
\alias{sn_active_reduction,Shennong-method}
\alias{sn_active_reduction<-,Shennong,character-method}
\title{Get the active reduction name for a given assay}
\usage{
sn_active_reduction(object, assay = NULL)

sn_active_reduction(object) <- value

\S4method{sn_active_reduction}{Shennong}(object, assay = NULL)

\S4method{sn_active_reduction}{Shennong,character}(object) <- value
}
\arguments{
\item{object}{A Shennong object.}

\item{assay}{Character name of the assay (optional).}

\item{value}{Name of the reduction to set as active.}
}
\value{
A character string.

The updated Shennong object.
}
\description{
Get the active reduction name for a given assay

Set the active reduction for an assay
}
