% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_add_assay}
\alias{sn_add_assay}
\alias{sn_add_assay,Shennong-method}
\title{Add an assay to a Shennong object}
\usage{
sn_add_assay(object, data, assay, layer = "counts")

\S4method{sn_add_assay}{Shennong}(object, data, assay, layer = "counts")
}
\arguments{
\item{object}{A Shennong object.}

\item{data}{A matrix-like object or SummarizedExperiment.}

\item{assay}{Character, name of the new assay.}

\item{layer}{Optional. Name of the data layer (default = "counts").}
}
\value{
A Shennong object with the new assay added.
}
\description{
Add an assay to a Shennong object
}
