% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_add_metadata}
\alias{sn_add_metadata}
\alias{AddMetaData}
\alias{sn_add_metadata,Shennong-method}
\title{Add Sample Metadata}
\usage{
sn_add_metadata(object, metadata, col_name = NULL, ...)

\S4method{sn_add_metadata}{Shennong}(object, metadata, col_name = NULL, ...)
}
\arguments{
\item{object}{A Shennong object.}

\item{metadata}{A data frame or DataFrame containing the new metadata columns.
Row names must match primary sample names in the object.}

\item{col_name}{Optional: If `metadata` is a single vector, specify the name
for the new metadata column.}

\item{...}{Ignored.}
}
\value{
The modified Shennong object.
}
\description{
Adds new columns to the main sample metadata table (`colData`).
}
