% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_assays}
\alias{sn_assays}
\alias{sn_assays,Shennong-method}
\title{Get Assay Names}
\usage{
sn_assays(object, ...)

\S4method{sn_assays}{Shennong}(object, ...)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{...}{Additional arguments.}
}
\description{
Get Assay Names
}
