% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-command.R
\name{sn_commands}
\alias{sn_commands}
\alias{sn_commands,Shennong-method}
\title{Get Command History}
\usage{
sn_commands(object, command = NULL, ...)

\S4method{sn_commands}{Shennong}(object, command = NULL, ...)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{command}{Optional: Specific command name.}

\item{...}{Additional arguments.}
}
\description{
Get Command History
}
\concept{command}
