% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{sn_create_reduction}
\alias{sn_create_reduction}
\title{Create a dimensional reduction object for Shennong}
\usage{
sn_create_reduction(
  embedding,
  loadings = matrix(0, nrow = 0, ncol = 0),
  stdev = numeric(),
  assay_used = NULL,
  key = NULL,
  global = FALSE,
  misc = list()
)
}
\arguments{
\item{embedding}{A matrix of sample coordinates in reduced space.}

\item{loadings}{Optional matrix of feature loadings.}

\item{stdev}{Optional numeric vector of standard deviations.}

\item{assay_used}{Assay name from which the reduction is derived.}

\item{key}{Character prefix used in column names of the embedding matrix.}

\item{global}{Logical. If TRUE, this reduction persists even if the assay is dropped.}

\item{misc}{List of additional metadata.}
}
\value{
A `ShennongReduction` object.
}
\description{
Create a dimensional reduction object for Shennong
}
