% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-shennong.R
\name{dim,Shennong-method}
\alias{dim,Shennong-method}
\title{Get Dimensions of Shennong Object}
\usage{
\S4method{dim}{Shennong}(x)
}
\arguments{
\item{x}{A Shennong object.}
}
\value{
A numeric vector of length 2: c(n_features, n_samples).
}
\description{
Returns the number of features in the active assay and the number of primary samples.
}
