% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-shennong.R
\name{dimnames,Shennong-method}
\alias{dimnames,Shennong-method}
\title{Get Dimension Names of Shennong Object}
\usage{
\S4method{dimnames}{Shennong}(x)
}
\arguments{
\item{x}{A Shennong object.}
}
\value{
A list of length 2 containing:
  1. Feature names (from active assay)
  2. Sample names (primary samples)
}
\description{
Returns the feature names from the active assay and the primary sample names.
}
