% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-reduction.R,
%   R/methods-shennong.R
\name{sn_embeddings}
\alias{sn_embeddings}
\alias{sn_embeddings,ShennongReduction-method}
\alias{sn_embeddings,Shennong-method}
\title{Get Embeddings}
\usage{
sn_embeddings(object, reduction = NULL, ...)

\S4method{sn_embeddings}{ShennongReduction}(object)

\S4method{sn_embeddings}{Shennong}(object, reduction = NULL)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{reduction}{The name of the reduction.}

\item{...}{Additional arguments.}
}
\description{
Get Embeddings
}
