% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_features}
\alias{sn_features}
\alias{sn_features,Shennong-method}
\title{Get Feature Names}
\usage{
sn_features(object, assay = NULL, ...)

\S4method{sn_features}{Shennong}(object, assay = NULL, ...)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{assay}{Optional: Assay context.}

\item{...}{Additional arguments.}
}
\description{
Get Feature Names
}
