% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-shennong.R
\name{head,Shennong-method}
\alias{head,Shennong-method}
\alias{head.Shennong}
\alias{tail.Shennong}
\alias{tail,Shennong-method}
\title{Head/Tail of Shennong object}
\usage{
\S4method{head}{Shennong}(x, n = 6L, ...)

\S4method{tail}{Shennong}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \code{Shennong} object.}

\item{n}{Number of rows to show. Default is 6.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input \code{Shennong} object.
}
\description{
Show the top or bottom rows of `colData()` in a `Shennong` object.
Useful for quick inspection of sample metadata.
}
