% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_layers}
\alias{sn_layers}
\alias{sn_layers,Shennong-method}
\title{Get Layer Names}
\usage{
sn_layers(object, assay = NULL, ...)

\S4method{sn_layers}{Shennong}(object, assay = NULL, ...)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{assay}{Optional: Assay name.}

\item{...}{Additional arguments.}
}
\description{
Get Layer Names
}
