% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_levels}
\alias{sn_levels}
\alias{sn_levels<-}
\alias{sn_levels,Shennong-method}
\alias{sn_levels<-,Shennong-method}
\title{Get Identity Levels}
\usage{
sn_levels(object, ...)

sn_levels(object, ...) <- value

\S4method{sn_levels}{Shennong}(object, ...)

\S4method{sn_levels}{Shennong}(object, ...) <- value
}
\arguments{
\item{object}{A Shennong object.}

\item{...}{Additional arguments.}

\item{value}{New levels (character vector).}
}
\description{
Get Identity Levels

Set Identity Levels<-
}
