% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-reduction.R,
%   R/methods-shennong.R
\name{sn_loadings}
\alias{sn_loadings}
\alias{sn_loadings,ShennongReduction-method}
\alias{sn_loadings,Shennong-method}
\title{Get Feature Loadings}
\usage{
sn_loadings(object, reduction = NULL, ...)

\S4method{sn_loadings}{ShennongReduction}(object)

\S4method{sn_loadings}{Shennong}(object, reduction = NULL)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{reduction}{The name of the reduction.}

\item{...}{Additional arguments.}
}
\description{
Get Feature Loadings
}
\concept{reduction}
