% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_metadata}
\alias{sn_metadata}
\alias{sn_metadata,Shennong-method}
\title{Get Sample Metadata}
\usage{
sn_metadata(object, cols = NULL, drop = FALSE, ...)

\S4method{sn_metadata}{Shennong}(object, cols = NULL, drop = FALSE, ...)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{cols}{Optional: Columns to retrieve.}

\item{drop}{Logical: Drop dimensions if possible.}

\item{...}{Additional arguments.}
}
\description{
Get Sample Metadata
}
