% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sn_misc}
\alias{sn_misc}
\alias{sn_misc<-}
\title{Get Miscellaneous Data}
\usage{
sn_misc(object, slot = NULL, ...)

sn_misc(object, slot, ...) <- value
}
\arguments{
\item{object}{A Shennong object.}

\item{slot}{Name for the misc data element.}

\item{...}{Additional arguments.}

\item{value}{The data to store.}
}
\description{
Get Miscellaneous Data

Set Miscellaneous Data<-
}
