% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_project_name}
\alias{sn_project_name}
\alias{sn_project_name<-}
\alias{sn_project_name,Shennong-method}
\alias{sn_project_name<-,Shennong-method}
\title{Get Project Name}
\usage{
sn_project_name(object, ...)

sn_project_name(object, ...) <- value

\S4method{sn_project_name}{Shennong}(object, ...)

\S4method{sn_project_name}{Shennong}(object, ...) <- value
}
\arguments{
\item{object}{A Shennong object.}

\item{...}{Additional arguments.}

\item{value}{New project name (character string).}
}
\description{
Get Project Name

Set Project Name<-
}
