% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sn_reductions}
\alias{sn_reductions}
\alias{sn_reductions<-}
\title{Get Dimensional Reductions}
\usage{
sn_reductions(object, reduction = NULL, ...)

sn_reductions(object, reduction, ...) <- value
}
\arguments{
\item{object}{A Shennong object.}

\item{reduction}{Name for the reduction.}

\item{...}{Additional arguments.}

\item{value}{The reduction data (list).}
}
\description{
Get Dimensional Reductions

Set Dimensional Reduction<-
}
