% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sn_reorder_ident}
\alias{sn_reorder_ident}
\title{Reorder Sample Identities}
\usage{
sn_reorder_ident(
  object,
  var,
  reverse = FALSE,
  fun = mean,
  reorder.numeric = FALSE,
  ...
)
}
\arguments{
\item{object}{A Shennong object.}

\item{var}{Variable to use for ordering.}

\item{reverse}{Reverse order.}

\item{fun}{Aggregation function.}

\item{reorder.numeric}{Rename levels numerically.}

\item{...}{Additional arguments.}
}
\description{
Reorder Sample Identities
}
