% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_samples}
\alias{sn_samples}
\alias{sn_samples,Shennong-method}
\title{Get Sample Names}
\usage{
sn_samples(object, assay = NULL, ...)

\S4method{sn_samples}{Shennong}(object, assay = NULL, ...)
}
\arguments{
\item{object}{A Shennong object or other class for which methods are defined.}

\item{assay}{Optional: Assay context.}

\item{...}{Additional arguments.}
}
\description{
Get Sample Names
}
