% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sn_samples_by_identities}
\alias{sn_samples_by_identities}
\title{Get samples grouped by identity class}
\usage{
sn_samples_by_identities(
  object,
  idents = NULL,
  samples = NULL,
  return.null = FALSE
)
}
\arguments{
\item{object}{A Shennong object.}

\item{idents}{Identity levels to include.}

\item{samples}{Samples to consider.}

\item{return.null}{Return NULL if no samples found.}
}
\description{
Get samples grouped by identity class
}
