% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sn_stash_ident}
\alias{sn_stash_ident}
\title{Stash Current Identities}
\usage{
sn_stash_ident(object, save.name = "orig.ident", ...)
}
\arguments{
\item{object}{A Shennong object.}

\item{save.name}{Column name to save identities under.}

\item{...}{Additional arguments.}
}
\description{
Stash Current Identities
}
