% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sn_stdev}
\alias{sn_stdev}
\title{Get Standard Deviations}
\usage{
sn_stdev(object, reduction, ...)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{reduction}{The name of the reduction.}

\item{...}{Additional arguments.}
}
\description{
Get Standard Deviations
}
