% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sn_tools}
\alias{sn_tools}
\alias{sn_tools<-}
\title{Get Tool Data}
\usage{
sn_tools(object, tool = NULL, ...)

sn_tools(object, tool, ...) <- value
}
\arguments{
\item{object}{A Shennong object.}

\item{tool}{Name of the tool.}

\item{...}{Additional arguments.}

\item{value}{The data generated by the tool.}
}
\description{
Get Tool Data

Set Tool Data<-
}
