% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-shennong.R
\name{sn_version}
\alias{sn_version}
\alias{sn_version,Shennong-method}
\title{Get Version Information}
\usage{
sn_version(object, ...)

\S4method{sn_version}{Shennong}(object, ...)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{...}{Additional arguments.}
}
\description{
Get Version Information
}
