% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sn_commands}
\alias{sn_commands}
\alias{sn_commands<-}
\title{Get Command History}
\usage{
sn_commands(object, command = NULL, ...)

sn_commands(object, command, ...) <- value
}
\arguments{
\item{object}{A Shennong object.}

\item{command}{Name for the command log entry.}

\item{...}{Additional arguments.}

\item{value}{The command log data.}
}
\description{
Get Command History

Set Command History<-
}
