% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sn_embeddings}
\alias{sn_embeddings}
\title{Get Embeddings}
\usage{
sn_embeddings(object, reduction, ...)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{reduction}{The name of the reduction.}

\item{...}{Additional arguments.}
}
\description{
Get Embeddings
}
