% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sn_loadings}
\alias{sn_loadings}
\title{Get Feature Loadings}
\usage{
sn_loadings(object, reduction, ...)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{reduction}{The name of the reduction.}

\item{...}{Additional arguments.}
}
\description{
Get Feature Loadings
}
