% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{sn_log_shennong_command}
\alias{sn_log_shennong_command}
\title{Log a Shennong command}
\usage{
sn_log_shennong_command(object, return_command = FALSE, .info = list())
}
\arguments{
\item{object}{A Shennong object}

\item{return_command}{Logical; return command object instead of storing it}

\item{.info}{A named list of parameters to inject into the command log}
}
\value{
A Shennong object or a ShennongCommand object
}
\description{
Captures the current function call and parameters for provenance tracking.
}
