% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-reduction.R
\name{sn_reductions}
\alias{sn_reductions}
\alias{sn_reductions,Shennong-method}
\title{Get Dimensional Reductions}
\usage{
sn_reductions(object, ...)

\S4method{sn_reductions}{Shennong}(object, ...)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{...}{Additional arguments.}
}
\description{
Get Dimensional Reductions
}
