% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-reduction.R,
%   R/methods-shennong.R
\name{sn_stdev}
\alias{sn_stdev}
\alias{sn_stdev,ShennongReduction-method}
\alias{sn_stdev,Shennong-method}
\title{Get Standard Deviations}
\usage{
sn_stdev(object, reduction = NULL, ...)

\S4method{sn_stdev}{ShennongReduction}(object)

\S4method{sn_stdev}{Shennong}(object, reduction = NULL)
}
\arguments{
\item{object}{A Shennong object or other class.}

\item{reduction}{The name of the reduction.}

\item{...}{Additional arguments.}
}
\description{
Get Standard Deviations
}
